Cris 06072023
Add two new models to the data base for users manipulation, one to store the possible preferences and another one to save the preferences for each user in the application.

In addition, add a new view for the user to select the initial preferences when the user create the profile for the first time as part of the user creation process.

@3thanF and @JeffRCH18 please be aware of the implementation of the new models on the DB, you will need to use the preferences tables as well for your applications.